# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive Electronic Medical Record system for hospitals",
        "title": "Hospital EMR System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@hospital-emr.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.hospital-emr.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/patients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of patients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "List patients",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new patient in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Create a new patient",
                "parameters": [
                    {
                        "description": "Patient information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_patient.CreatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/patients/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Get patient by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update patient information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Update patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_patient.CreatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft delete a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Delete patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/v1/patients/{id}/timeline": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get complete medical timeline for a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Get patient timeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_hospital-emr_backend_internal_common_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_hospital-emr_backend_internal_common_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "github_com_hospital-emr_backend_internal_models.Allergy": {
            "type": "object",
            "properties": {
                "allergen": {
                    "type": "string"
                },
                "allergy_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AllergyType"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "onset_date": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "reaction": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AllergySeverity"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "github_com_hospital-emr_backend_internal_models.AllergySeverity": {
            "type": "string",
            "enum": [
                "mild",
                "moderate",
                "severe",
                "fatal"
            ],
            "x-enum-varnames": [
                "AllergySeverityMild",
                "AllergySeverityModerate",
                "AllergySeveritySevere",
                "AllergySeverityFatal"
            ]
        },
        "github_com_hospital-emr_backend_internal_models.AllergyType": {
            "type": "string",
            "enum": [
                "drug",
                "food",
                "environment",
                "other"
            ],
            "x-enum-varnames": [
                "AllergyTypeDrug",
                "AllergyTypeFood",
                "AllergyTypeEnvironment",
                "AllergyTypeOther"
            ]
        },
        "github_com_hospital-emr_backend_internal_models.Appointment": {
            "type": "object",
            "properties": {
                "appointment_number": {
                    "type": "string"
                },
                "appointment_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AppointmentType"
                },
                "cancellation_reason": {
                    "type": "string"
                },
                "cancelled_at": {
                    "type": "string"
                },
                "checked_in_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "type": "string"
                },
                "duration": {
                    "description": "in minutes",
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "patient": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                },
                "patient_id": {
                    "type": "string"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.User"
                },
                "provider_id": {
                    "type": "string"
                },
                "reason_for_visit": {
                    "type": "string"
                },
                "reminder_sent": {
                    "type": "boolean"
                },
                "reminder_sent_at": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.AppointmentStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.EmergencyContact": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "relationship": {
                    "type": "string"
                }
            }
        },
        "github_com_hospital-emr_backend_internal_models.Encounter": {
            "type": "object",
            "properties": {
                "admission_date": {
                    "type": "string"
                },
                "chief_complaint": {
                    "type": "string"
                },
                "clinical_notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.ClinicalNote"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "department": {
                    "type": "string"
                },
                "diagnoses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Diagnosis"
                    }
                },
                "discharge_date": {
                    "type": "string"
                },
                "encounter_number": {
                    "type": "string"
                },
                "encounter_type": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EncounterType"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Order"
                    }
                },
                "patient": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Patient"
                },
                "patient_id": {
                    "type": "string"
                },
                "priority": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Priority"
                },
                "procedures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Procedure"
                    }
                },
                "provider": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.User"
                },
                "provider_id": {
                    "type": "string"
                },
                "reason_for_visit": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EncounterStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "vital_signs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.VitalSign"
                    }
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "other",
                "unknown"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderOther",
                "GenderUnknown"
            ]
        },
        "github_com_hospital-emr_backend_internal_models.Insurance": {
            "type": "object",
            "properties": {
                "coverage_type": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "group_number": {
                    "type": "string"
                },
                "policy_number": {
                    "type": "string"
                },
                "primary_insured": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "github_com_hospital-emr_backend_internal_models.MaritalStatus": {
            "type": "string",
            "enum": [
                "single",
                "married",
                "divorced",
                "widowed"
            ],
            "x-enum-varnames": [
                "MaritalStatusSingle",
                "MaritalStatusMarried",
                "MaritalStatusDivorced",
                "MaritalStatusWidowed"
            ]
        },
...
        "github_com_hospital-emr_backend_internal_models.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "allergies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Allergy"
                    }
                },
                "appointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Appointment"
                    }
                },
                "blood_type": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EmergencyContact"
                },
                "encounters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Encounter"
                    }
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Gender"
                },
                "id": {
                    "type": "string"
                },
                "insurance": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Insurance"
                },
                "language": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "marital_status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.MaritalStatus"
                },
                "medications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Medication"
                    }
                },
                "middle_name": {
                    "type": "string"
                },
                "mobile_number": {
                    "type": "string"
                },
                "mrn": {
                    "description": "Medical Record Number",
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "profile_photo": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "ssn": {
                    "description": "Social Security Number / National ID",
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.PatientStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
        "github_com_hospital-emr_backend_internal_models.PatientStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "deceased"
            ],
            "x-enum-varnames": [
                "PatientStatusActive",
                "PatientStatusInactive",
                "PatientStatusDeceased"
            ]
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_patient.CreatePatientRequest": {
            "type": "object",
            "required": [
                "date_of_birth",
                "first_name",
                "gender",
                "last_name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "blood_type": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.EmergencyContact"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Gender"
                },
                "insurance": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.Insurance"
                },
                "language": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "marital_status": {
                    "$ref": "#/definitions/github_com_hospital-emr_backend_internal_models.MaritalStatus"
                },
                "middle_name": {
                    "type": "string"
                },
                "mobile_number": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "ssn": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}